JobAPIService/requirements.txt
flask
requests
redis
PyJWT